Moved calculator namespace
